requirements.txt
python-telegram-bot==20.3
APScheduler==3.10.4
pytz==2024.1
